Adding music library with .strm file redirection
Added music library files through a redirection in .strm files instead
of a direct link to the plex server (giving issues with music scraper)
or direct link to plugin (not allowed by kodi)

Basically; to the same as for movies but now through .strm files which
are saved in the addon_data folder.
Though, this wasn't a simple task. First of all the scraper was deleting
files if it couldn't verify their integrity (which it couldn't because a
.strm is not a music file for the scraper). Therefore the directory is
not told to the scraper, but instead the absolute path is parsed as
filename and the dir set to ''.
Because this path is usually followed for video files, some breakout
were installed for correct function.